requirement.txt
dcard-spider==0.3.0
snownlp
matplotlib==2.2.2
pymongo==3.5.1
emoji==0.5.3
seaborn==0.9.0
pyshorteners==0.6.1
monthdelta
imgurpython
selenium==3.141.0
wordcloud==1.5.0
jieba==0.39
opencc-python-reimplemented==0.1.4
twstock==1.3.1
lxml==4.4.1
flask-bcrypt==0.7.1
bs4
line-bot-sdk
